+++ src/Client/Auth/NoAuth.php
@@ -11,6 +11,8 @@ use Addwiki\Mediawiki\Api\Client\Request\Request;
 class NoAuth implements AuthMethod {
 
 	public function preRequestAuth( string $method, Request $request, MediawikiApi $api ): Request {
+		// Verify that the user is logged in if set to user, not logged in if set to anon, or has the bot user right if bot.
+		$request->setParam( 'assert', 'anon' );
 		return $request;
 	}