+++ src/Client/Auth/OAuthOwnerConsumer.php
@@ -54,6 +54,9 @@ class OAuthOwnerConsumer implements AuthMethod {
 	}
 
 	public function preRequestAuth( string $method, Request $request, MediawikiApi $api ): Request {
+		// Verify that the user is logged in if set to user, not logged in if set to anon, or has the bot user right if bot.
+		$request->setParam( 'assert', 'user' );
+
 		$request->setHeaders( array_merge( $request->getHeaders(), [ 'Authorization' => $this->getAuthenticationHeaderValue( $method, $request, $api ) ] ) );
 		return $request;
 	}