+++ tests/unit/Client/Auth/UserAndPasswordTest.php
@@ -3,8 +3,13 @@
 namespace Addwiki\Mediawiki\Api\Tests\Unit\Client\Auth;
 
 use Addwiki\Mediawiki\Api\Client\Auth\UserAndPassword;
+use Addwiki\Mediawiki\Api\Client\MediawikiApi;
+use Addwiki\Mediawiki\Api\Client\Request\SimpleRequest;
+use Addwiki\Mediawiki\Api\Client\UsageException;
+use GuzzleHttp\ClientInterface;
 use InvalidArgumentException;
 use PHPUnit\Framework\TestCase;
+use Psr\Http\Message\ResponseInterface;
 
 /**
  * @covers Mediawiki\Api\Client\Auth\UserAndPassword
@@ -58,6 +63,84 @@ class UserAndPasswordTest extends TestCase {
 			[ new UserAndPassword( 'usera', 'passa' ), new UserAndPassword( 'DIFF', 'passa' ), false ],
 			[ new UserAndPassword( 'usera', 'passa' ), new UserAndPassword( 'usera', 'DIFF' ), false ],
 		];
+	}
+
+	/**
+	 * @return MockObject&ResponseInterface
+	 */
+	private function getMockResponse( $responseValue ) {
+		$mock = $this->createMock( ResponseInterface::class );
+		$mock
+			->method( 'getBody' )
+			->willReturn( json_encode( $responseValue ) );
+		return $mock;
+	}
+
+	/**
+	 * @return array <int|string mixed[]>
+	 */
+	private function getExpectedRequestOpts( $params, $paramsLocation ): array {
+		return [
+			$paramsLocation => array_merge( $params, [ 'format' => 'json' ] ),
+			'headers' => [ 'User-Agent' => 'addwiki-mediawiki-client/U1' ],
+		];
+	}
+
+	public function testGoodLoginSequence(): void {
+		$eq1 = [
+			'action' => 'login',
+			'lgname' => 'U1',
+			'lgpassword' => 'P1',
+		];
+		$params = array_merge( $eq1, [ 'lgtoken' => 'IamLoginTK' ] );
+
+		$client = $this->createMock( ClientInterface::class );
+		$client->expects( $this->exactly( 2 ) )
+			->method( 'request' )
+			->withConsecutive(
+				[ 'POST', null, $this->getExpectedRequestOpts( $eq1, 'form_params' ) ],
+				[ 'POST', null, $this->getExpectedRequestOpts( $params, 'form_params' ) ],
+			)
+			->willReturnOnConsecutiveCalls(
+				$this->getMockResponse( [ 'login' => [
+					'result' => 'NeedToken',
+					'token' => 'IamLoginTK',
+				] ] ),
+				$this->getMockResponse( [ 'login' => [ 'result' => 'Success' ] ] )
+			);
+
+		$auth = new UserAndPassword( 'U1', 'P1' );
+		$api = new MediawikiApi( '', $auth, $client );
+		$auth->preRequestAuth( 'GET', new SimpleRequest( 'dummyrequest' ), $api );
+	}
+
+	public function testBadLoginSequence(): void {
+		$client = $this->createMock( ClientInterface::class );
+		$eq1 = [
+			'action' => 'login',
+			'lgname' => 'U1',
+			'lgpassword' => 'P1',
+		];
+		$params = array_merge( $eq1, [ 'lgtoken' => 'IamLoginTK' ] );
+
+		$client->expects( $this->exactly( 2 ) )
+			->method( 'request' )
+			->withConsecutive(
+				[ 'POST', null, $this->getExpectedRequestOpts( $eq1, 'form_params' ) ],
+				[ 'POST', null, $this->getExpectedRequestOpts( $params, 'form_params' ) ],
+			)
+			->willReturnOnConsecutiveCalls(
+				$this->getMockResponse( [ 'login' => [
+					'result' => 'NeedToken',
+					'token' => 'IamLoginTK',
+				] ] ),
+				$this->getMockResponse( [ 'login' => [ 'result' => 'BADTOKENorsmthin' ] ] )
+			);
+
+		$auth = new UserAndPassword( 'U1', 'P1' );
+		$api = new MediawikiApi( '', $auth, $client );
+		$this->expectException( UsageException::class );
+		$auth->preRequestAuth( 'GET', new SimpleRequest( 'dummyrequest' ), $api );
 	}
 
 }